+++ module/documents/active-effect/torgActiveEffect.js
@@ -1,3 +1,5 @@
+import { TestResult } from '../../torgchecks.js';
+
 /**
  * Extend the basic ActiveEffect model with migrations and TORG specific handling
  */
@@ -79,4 +81,29 @@ export default class TorgActiveEffect extends foundry.documents.ActiveEffect {
     if (!this.parent || this.parent instanceof Actor) return game.i18n.localize("None");
     return this.parent.name;
   }
+
+  /**
+   * 
+   * @param {TestResult} result 
+   * @param {Array<String> | undefined} attackTraits array of traits on the actor performing the test
+   * @param {Array<String> | undefined} defendTraits array of traits on the target of the test
+   */
+  appliesToTest(result, attackTraits, defendTraits) {
+    return ((this.system.transferOnAttack && result >= TestResult.STANDARD) || (this.system.transferOnOutcome === result)) &&
+      testTraits(this.system.applyIfAttackTrait, attackTraits) &&
+      testTraits(this.system.applyIfDefendTrait, defendTraits);
+  }
+}
+
+/**
+ * Return true if testTraits contains at least one of the entries in actualTraits
+ * @param {Set<String>} testTraits 
+ * @param {Array<String>} actualTraits 
+ */
+function testTraits(testTraits, actualTraits) {
+  if (!testTraits?.size) return true;
+  for (const trait of testTraits) {
+    if (actualTraits.includes(trait)) return true;
+  }
+  return false;
 }