+++ Changelog.md
@@ -5,6 +5,8 @@
 - Get Active Effect "Transfer on Attack" flag working (as part of making our own Data Model for Active Effects).
 - Fix "Cosm Possibilities" dialog working again.
 - Fix styling of journal entries (especially image subtitles and table headers).
+- Prevent double dice rolling with dice-so-nice when only a single (or no) target is selected.
+- **Active Effects** on an item can conditionally be applied to the target based on the attack/defense traits.
 
 ## 13.10.0